Selectors refactored
