+++ src/components/NotesPage/NotesPageComponent.js
@@ -5,7 +5,7 @@ import { ControlsHelpComponent } from 'components/ControlsHelp/ControlsHelpCompo
 import { NotesContentEditorContainer } from 'components/NotesContentEditor/NotesContentEditorContainer';
 import { NotesMindMapContainer } from 'components/NotesMindMap/NotesMindMapContainer';
 import { SearchPanelContainer } from 'components/SearchPanel/SearchPanelContainer';
-import { StatusButtonComponent } from 'components/Uploads/StatusButtonComponent';
+import { StatusButtonComponent } from 'components/Uploads/UploadsStatusButtonComponent';
 import { UploadButtonComponent } from 'components/Uploads/UploadButtonComponent';
 
 export class NotesPageComponent extends Component {
